+++ packages/activity_recognition_flutter/lib/activity_recognition_flutter.dart
@@ -5,14 +5,28 @@ import 'dart:convert';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
-part 'package:activity_recognition_flutter/channel/activity_channel.dart';
-part 'package:activity_recognition_flutter/data/activity.dart';
+import 'package:flutter_foreground_service/flutter_foreground_service.dart';
+import 'dart:io' show Platform;
+
+part 'package:activity_recognition_flutter/src/ar_channel.dart';
+
+part 'package:activity_recognition_flutter/src/ar_domain.dart';
 
 class ActivityRecognition {
-  /// Requests continuous [Activity] updates.
-  ///
-  /// The Stream will output the *most probable* [Activity].
-  static Stream<Activity> activityUpdates() => _activityChannel.activityUpdates;
+  static _ActivityChannel _channel;
 
-  static final _activityChannel = _ActivityChannel();
+  /// Requests continuous [ActivityEvent] updates.
+  ///
+  /// The Stream will output the *most probable* [ActivityEvent].
+  /// By default the foreground service is enabled, which allows the
+  /// updates to be streamed while the app runs in the background.
+  /// The programmer can choose to not enable to foreground service,
+  /// if they so choose.
+  static Stream<ActivityEvent> activityStream(
+      {bool runForegroundService = true}) {
+    if (_channel == null) {
+      _channel = _ActivityChannel(runForegroundService);
+    }
+    return _channel.activityUpdates;
+  }
 }