+++ packages/app_usage/example/lib/main.dart
@@ -18,12 +18,16 @@ class _MyAppState extends State<MyApp> {
 
   void getUsageStats() async {
     try {
-      DateTime startDate = DateTime(2018, 01, 01);
       DateTime endDate = new DateTime.now();
-      List<AppUsageInfo> infos = await AppUsage.getAppUsage(startDate, endDate);
+      DateTime startDate = endDate.subtract(Duration(hours: 1));
+      List<AppUsageInfo> infoList = await AppUsage.getAppUsage(startDate, endDate);
       setState(() {
-        _infos = infos;
+        _infos = infoList;
       });
+
+      for (var info in infoList) {
+        print(info.toString());
+      }
     } on AppUsageException catch (exception) {
       print(exception);
     }