+++ packages/esense_flutter/example/android/app/src/main/java/dk/cachet/esense_flutter_example/MainActivity.java
@@ -1,13 +1,6 @@
 package dk.cachet.esense_flutter_example;
 
-import android.os.Bundle;
-import io.flutter.app.FlutterActivity;
-import io.flutter.plugins.GeneratedPluginRegistrant;
+import io.flutter.embedding.android.FlutterActivity;
 
 public class MainActivity extends FlutterActivity {
-  @Override
-  protected void onCreate(Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-    GeneratedPluginRegistrant.registerWith(this);
-  }
 }