+++ packages/esense_flutter/example/lib/main.dart
@@ -1,7 +1,5 @@
 import 'package:flutter/material.dart';
 import 'dart:async';
-
-import 'package:flutter/services.dart';
 import 'package:esense_flutter/esense.dart';
 
 void main() => runApp(MyApp());
@@ -18,30 +16,31 @@ class _MyAppState extends State<MyApp> {
   bool sampling = false;
   String _event = '';
   String _button = 'not pressed';
+  bool connected = false;
 
   // the name of the eSense device to connect to -- change this to your own device.
   String eSenseName = 'eSense-0332';
 
-  @override
   void initState() {
     super.initState();
-    _connectToESense();
+    _listenToESense();
   }
 
-  Future<void> _connectToESense() async {
-    bool con = false;
-
-    // if you want to get the connection events when connecting, set up the listener BEFORE connecting...
-    ESenseManager.connectionEvents.listen((event) {
+  Future _listenToESense() async {
+    // if you want to get the connection events when connecting,
+    // set up the listener BEFORE connecting...
+    ESenseManager().connectionEvents.listen((event) {
       print('CONNECTION event: $event');
 
       // when we're connected to the eSense device, we can start listening to events from it
       if (event.type == ConnectionType.connected) _listenToESenseEvents();
 
       setState(() {
+        connected = false;
         switch (event.type) {
           case ConnectionType.connected:
             _deviceStatus = 'connected';
+            connected = true;
             break;
           case ConnectionType.unknown:
             _deviceStatus = 'unknown';
@@ -58,16 +57,19 @@ class _MyAppState extends State<MyApp> {
         }
       });
     });
+  }
 
-    con = await ESenseManager.connect(eSenseName);
+  Future _connectToESense() async {
+    print('connecting... connected: $connected');
+    if (!connected) connected = await ESenseManager().connect(eSenseName);
 
     setState(() {
-      _deviceStatus = con ? 'connecting' : 'connection failed';
+      _deviceStatus = connected ? 'connecting' : 'connection failed';
     });
   }
 
   void _listenToESenseEvents() async {
-    ESenseManager.eSenseEvents.listen((event) {
+    ESenseManager().eSenseEvents.listen((event) {
       print('ESENSE event: $event');
 
       setState(() {
@@ -79,7 +81,9 @@ class _MyAppState extends State<MyApp> {
             _voltage = (event as BatteryRead).voltage;
             break;
           case ButtonEventChanged:
-            _button = (event as ButtonEventChanged).pressed ? 'pressed' : 'not pressed';
+            _button = (event as ButtonEventChanged).pressed
+                ? 'pressed'
+                : 'not pressed';
             break;
           case AccelerometerOffsetRead:
             // TODO
@@ -99,21 +103,31 @@ class _MyAppState extends State<MyApp> {
 
   void _getESenseProperties() async {
     // get the battery level every 10 secs
-    Timer.periodic(Duration(seconds: 10), (timer) async => await ESenseManager.getBatteryVoltage());
+    Timer.periodic(
+      Duration(seconds: 10),
+      (timer) async =>
+          (connected) ? await ESenseManager().getBatteryVoltage() : null,
+    );
 
     // wait 2, 3, 4, 5, ... secs before getting the name, offset, etc.
     // it seems like the eSense BTLE interface does NOT like to get called
     // several times in a row -- hence, delays are added in the following calls
-    Timer(Duration(seconds: 2), () async => await ESenseManager.getDeviceName());
-    Timer(Duration(seconds: 3), () async => await ESenseManager.getAccelerometerOffset());
-    Timer(Duration(seconds: 4), () async => await ESenseManager.getAdvertisementAndConnectionInterval());
-    Timer(Duration(seconds: 5), () async => await ESenseManager.getSensorConfig());
+    Timer(Duration(seconds: 2),
+        () async => await ESenseManager().getDeviceName());
+    Timer(Duration(seconds: 3),
+        () async => await ESenseManager().getAccelerometerOffset());
+    Timer(
+        Duration(seconds: 4),
+        () async =>
+            await ESenseManager().getAdvertisementAndConnectionInterval());
+    Timer(Duration(seconds: 5),
+        () async => await ESenseManager().getSensorConfig());
   }
 
   StreamSubscription subscription;
   void _startListenToSensorEvents() async {
     // subscribe to sensor event from the eSense device
-    subscription = ESenseManager.sensorEvents.listen((event) {
+    subscription = ESenseManager().sensorEvents.listen((event) {
       print('SENSOR event: $event');
       setState(() {
         _event = event.toString();
@@ -133,7 +147,7 @@ class _MyAppState extends State<MyApp> {
 
   void dispose() {
     _pauseListenToSensorEvents();
-    ESenseManager.disconnect();
+    ESenseManager().disconnect();
     super.dispose();
   }
 
@@ -153,14 +167,31 @@ class _MyAppState extends State<MyApp> {
               Text('eSense Button Event: \t$_button'),
               Text(''),
               Text('$_event'),
+              Container(
+                height: 80,
+                width: 200,
+                decoration:
+                    BoxDecoration(borderRadius: BorderRadius.circular(10)),
+                child: TextButton.icon(
+                  onPressed: _connectToESense,
+                  icon: Icon(Icons.login),
+                  label: Text(
+                    'CONNECT....',
+                    style: TextStyle(fontSize: 35),
+                  ),
+                ),
+              ),
             ],
           ),
         ),
-        floatingActionButton: new FloatingActionButton(
+        floatingActionButton: FloatingActionButton(
           // a floating button that starts/stops listening to sensor events.
           // is disabled until we're connected to the device.
-          onPressed:
-              (!ESenseManager.connected) ? null : (!sampling) ? _startListenToSensorEvents : _pauseListenToSensorEvents,
+          onPressed: (!ESenseManager().connected)
+              ? null
+              : (!sampling)
+                  ? _startListenToSensorEvents
+                  : _pauseListenToSensorEvents,
           tooltip: 'Listen to eSense sensors',
           child: (!sampling) ? Icon(Icons.play_arrow) : Icon(Icons.pause),
         ),