+++ packages/flutter_foreground_service/example/android/app/src/main/kotlin/dk/cachet/flutter_foreground_service_example/MainActivity.kt
@@ -1,0 +1,6 @@
+package dk.cachet.flutter_foreground_service_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}