+++ packages/health/example/lib/main.dart
@@ -9,7 +9,13 @@ class MyApp extends StatefulWidget {
   _MyAppState createState() => _MyAppState();
 }
 
-enum AppState { DATA_NOT_FETCHED, FETCHING_DATA, DATA_READY, NO_DATA }
+enum AppState {
+  DATA_NOT_FETCHED,
+  FETCHING_DATA,
+  DATA_READY,
+  NO_DATA,
+  AUTH_NOT_GRANTED
+}
 
 class _MyAppState extends State<MyApp> {
   List<HealthDataPoint> _healthDataList = [];
@@ -21,41 +27,60 @@ class _MyAppState extends State<MyApp> {
   }
 
   Future<void> fetchData() async {
-    setState(() {
-      _state = AppState.FETCHING_DATA;
-    });
-
     /// Get everything from midnight until now
-    DateTime endDate = DateTime.now();
-    DateTime startDate = DateTime(2020, 01, 01);
+    DateTime startDate = DateTime(2020, 11,   07, 0,  0,  0);
+    DateTime endDate = DateTime(2020,   11,   07, 23, 59, 59);
 
     HealthFactory health = HealthFactory();
 
     /// Define the types to get.
     List<HealthDataType> types = [
-      HealthDataType.BODY_MASS_INDEX,
       HealthDataType.STEPS,
       HealthDataType.WEIGHT,
-      HealthDataType.ACTIVE_ENERGY_BURNED
+      HealthDataType.HEIGHT,
+      HealthDataType.BLOOD_GLUCOSE,
+      HealthDataType.DISTANCE_WALKING_RUNNING,
     ];
 
-    /// Fetch new data
-    List<HealthDataPoint> healthData =
-        await health.getHealthDataFromTypes(startDate, endDate, types);
+    setState(() => _state = AppState.FETCHING_DATA);
 
-    /// Save all the new data points
-    _healthDataList.addAll(healthData);
+    /// You MUST request access to the data types before reading them
+    bool accessWasGranted = await health.requestAuthorization(types);
 
-    /// Filter out duplicates
-    _healthDataList = HealthFactory.removeDuplicates(_healthDataList);
+    int steps = 0;
 
-    /// Print the results
-    _healthDataList.forEach((x) => print("Data point: $x"));
+    if (accessWasGranted) {
+      try {
+        /// Fetch new data
+        List<HealthDataPoint> healthData =
+            await health.getHealthDataFromTypes(startDate, endDate, types);
 
-    /// Update the UI to display the results
-    setState(() {
-      _state = _healthDataList.isEmpty ? AppState.NO_DATA : AppState.DATA_READY;
-    });
+        /// Save all the new data points
+        _healthDataList.addAll(healthData);
+      } catch (e) {
+        print("Caught exception in getHealthDataFromTypes: $e");
+      }
+
+      /// Filter out duplicates
+      _healthDataList = HealthFactory.removeDuplicates(_healthDataList);
+
+      /// Print the results
+      _healthDataList.forEach((x) {
+        print("Data point: $x");
+        steps += (x.value as int);
+      });
+
+      print("Steps: $steps");
+
+      /// Update the UI to display the results
+      setState(() {
+        _state =
+            _healthDataList.isEmpty ? AppState.NO_DATA : AppState.DATA_READY;
+      });
+    } else {
+      print("Authorization not granted");
+      setState(() => _state = AppState.DATA_NOT_FETCHED);
+    }
   }
 
   Widget _contentFetchingData() {
@@ -93,12 +118,21 @@ class _MyAppState extends State<MyApp> {
     return Text('Press the download button to fetch data');
   }
 
+  Widget _authorizationNotGranted() {
+    return Text('''Authorization not given.
+        For Android please check your OAUTH2 client ID is correct in Google Developer Console.
+         For iOS check your permissions in Apple Health.''');
+  }
+
   Widget _content() {
     if (_state == AppState.DATA_READY)
       return _contentDataReady();
     else if (_state == AppState.NO_DATA)
       return _contentNoData();
-    else if (_state == AppState.FETCHING_DATA) return _contentFetchingData();
+    else if (_state == AppState.FETCHING_DATA)
+      return _contentFetchingData();
+    else if (_state == AppState.AUTH_NOT_GRANTED)
+      return _authorizationNotGranted();
 
     return _contentNotFetched();
   }