+++ packages/health/lib/health.dart
@@ -1,13 +1,16 @@
 library health;
 
 import 'dart:async';
-import 'dart:io' show Platform;
-
-import 'package:device_id/device_id.dart';
 import 'package:flutter/services.dart';
+import 'dart:io' show Platform;
+import 'package:device_info/device_info.dart';
+
+part 'package:health/src/units.dart';
 
 part 'package:health/src/data_types.dart';
+
 part 'package:health/src/functions.dart';
+
 part 'package:health/src/health_data_point.dart';
+
 part 'package:health/src/health_factory.dart';
-part 'package:health/src/units.dart';