+++ packages/health/lib/src/units.dart
@@ -18,5 +18,6 @@ enum HealthDataUnit {
   NO_UNIT,
   PERCENTAGE,
   SIEMENS,
-  UNKNOWN_UNIT
+  UNKNOWN_UNIT,
+  LITER,
 }