+++ packages/activity_recognition_flutter/CHANGELOG.md
@@ -1,3 +1,17 @@
+## 2.0.2
+* upgrade to `flutter_foreground_service` v. 0.2.1 ([Issue #238](https://github.com/cph-cachet/flutter-plugins/issues/238))
+* updated wrong title in example app
+
+## 2.0.1
+* upgrade to `flutter_foreground_service` v. 0.2.0 ([Issue #238](https://github.com/cph-cachet/flutter-plugins/issues/238))
+
+## 2.0.0
+* Activity recognition now works in the background on Android by means of a foreground service
+* Updated domain model
+
+## 1.2.4
+* Updated example wrt. permission for Android API 29 and later
+
 ## 1.2.3
 * Added the ActivityType class
 * Updated example app