+++ packages/noise_meter/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.0.0
+* Changed API to support audio_streamer version 1.3.0
+* Changed documentation
+
 ## 1.2.0
 * Changed to api of a NoiseReading to include max and mean decibel.